Move WorldMovement.covert/1
Moves this function to the Move modules, because it only operates on a Move
struct.

Renames the function to direction_to_point/1
