+++ lib/battle_snake/world_movement.ex
@@ -25,7 +25,7 @@ defmodule BattleSnake.WorldMovement do
     update_in(world.snakes, fn snakes ->
       for snake <- snakes do
         move = moves[snake.id]
-        point = World.convert(move.move)
+        point = Move.direction_to_point(move.move)
         Snake.move(snake, point)
       end
     end)
